+++ src/main/java/org/hv/pocket/session/actions/OperateDictionary.java
@@ -30,7 +30,7 @@ public interface OperateDictionary {
 
 
     /**
-     * 查询对象
+     * 级联查询对象
      *
      * @param clazz    类类型
      * @param identify 数据标识
@@ -38,6 +38,18 @@ public interface OperateDictionary {
      * @throws SQLException e
      */
     <T extends AbstractEntity> T findOne(Class<T> clazz, Serializable identify) throws SQLException;
+
+
+    /**
+     * 查询对象
+     *
+     * @param clazz    类类型
+     * @param identify 数据标识
+     * @param cascade  是否进行级联保存操作
+     * @return 实体对象
+     * @throws SQLException e
+     */
+    <T extends AbstractEntity> T findOne(Class<T> clazz, Serializable identify, boolean cascade) throws SQLException;
 
     /**
      * 查询所有
@@ -57,7 +69,7 @@ public interface OperateDictionary {
     <E extends AbstractEntity> List<E> list(Class<E> clazz, boolean cascade);
 
     /**
-     * 强制通过数据库查询数据
+     * 强制通过数据库级联查询数据
      *
      * @param clazz    类类型
      * @param identify 数据标识
@@ -65,6 +77,17 @@ public interface OperateDictionary {
      * @throws SQLException 语句异常
      */
     <T extends AbstractEntity> T findDirect(Class<T> clazz, Serializable identify) throws SQLException;
+
+    /**
+     * 强制通过数据库查询数据
+     *
+     * @param clazz    类类型
+     * @param identify 数据标识
+     * @param cascade  是否进行级联保存操作
+     * @return 实体对象
+     * @throws SQLException 语句异常
+     */
+    <T extends AbstractEntity> T findDirect(Class<T> clazz, Serializable identify, boolean cascade) throws SQLException;
 
     /**
      * 保存(NULL不纳入保存范围保留数据库默认值)